Shifted tree donation
