+++ src/Donations/Components/DonationsForm.tsx
@@ -1,10 +1,8 @@
 import React from "react";
-import CustomIcon from "../../../public/assets/icons/CustomIcon";
 import { QueryParamContext } from "../../Layout/QueryParamContext";
 import GiftForm from "../Micros/GiftForm";
 import { useTranslation } from "next-i18next";
 import getFormatedCurrency from "../../Utils/getFormattedCurrency";
-import DownArrowIcon from "../../../public/assets/icons/DownArrowIcon";
 import { getMinimumAmountForCurrency } from "../../Utils/getExchange";
 import { formatAmountForStripe } from "../../Utils/stripe/stripeHelpers";
 import { NativePay } from "../PaymentMethods/PaymentRequestCustomButton";
@@ -14,15 +12,12 @@ import {
   payDonationFunction,
 } from "../PaymentMethods/PaymentFunctions";
 import PaymentProgress from "../../Common/ContentLoaders/Donations/PaymentProgress";
-import { getFormattedNumber } from "../../Utils/getFormattedNumber";
-import themeProperties from "../../../styles/themeProperties";
 import SelectCurrencyModal from "./../Micros/SelectCurrencyModal";
 import TaxDeductionOption from "./../Micros/TaxDeductionOption";
-import TreeCostLoader from "../../Common/ContentLoaders/TreeCostLoader";
 import Authentication from "./../Micros/Authentication";
 import { useAuth0 } from "@auth0/auth0-react";
-import { getCountryDataBy } from "../../Utils/countryUtils";
-
+import DonationAmount from "../Micros/DonationAmount";
+import TreeDonation from "../Micros/DonationTypes/TreeDonation"
 function DonationsForm() {
   const {
     isGift,
@@ -116,32 +111,13 @@ function DonationsForm() {
     });
   };
 
-  const [customTreeInputValue, setCustomTreeInputValue] = React.useState("");
 
-  const [isCustomDonation, setisCustomDonation] = React.useState(false);
 
-  const setCustomTreeValue = (e: any) => {
-    if (e.target) {
-      if (e.target.value === "" || e.target.value < 1) {
-        // if input is '', default 1
-        settreeCount(1);
-      } else if (e.target.value.toString().length <= 12) {
-        settreeCount(e.target.value);
-      }
-    }
-  };
 
-  React.useEffect(() => {
-    if (![10, 20, 50, 150].includes(treeCount)) {
-      setisCustomDonation(true);
-      setCustomTreeValue(treeCount);
-      setCustomTreeInputValue(treeCount);
-    }
-  }, [treeCount]);
 
   const [openCurrencyModal, setopenCurrencyModal] = React.useState(false);
 
-  const customInputRef = React.useRef(null);
+
 
   return isPaymentProcessing ? (
     <PaymentProgress isPaymentProcessing={isPaymentProcessing} />
@@ -154,165 +130,57 @@ function DonationsForm() {
           <div className="donations-gift-container mt-10">
             <GiftForm />
           </div>
-          <div
-            className={`donations-tree-selection ${
-              isGift && giftDetails.recipientName === "" ? "display-none" : ""
-            }`}
-          >
-            <div className="tree-selection-options-container">
-              {treeSelectionOptions.map((option, key) => {
-                return (
-                  <div
-                    onClick={() => {
-                      settreeCount(option.treeCount);
-                      setisCustomDonation(false);
-                      setCustomTreeInputValue("");
-                    }}
-                    key={key}
-                    className={`tree-selection-option mt-20 ${
-                      option.treeCount === treeCount && !isCustomDonation
-                        ? "tree-selection-option-selected"
-                        : ""
-                    }`}
-                  >
-                    {option.iconFile}
-                    <div className="tree-selection-option-text">
-                      <p>{option.treeCount}</p>
-                      <span>{t("trees")}</span>
-                    </div>
-                  </div>
-                );
-              })}
+          
+          <TreeDonation setopenCurrencyModal={setopenCurrencyModal} />
 
-              <div
-                className={`tree-selection-option mt-20 custom-selection ${
-                  isCustomDonation ? "tree-selection-option-selected" : ""
-                }`}
-                onClick={() => {
-                  setisCustomDonation(true);
-                  customInputRef.current.focus();
-                }}
-              >
-                <CustomIcon />
-                <div className="tree-selection-option-text">
-                  <input
-                    className={"custom-tree-input"}
-                    onInput={(e) => {
-                      // replaces any character other than number to blank
-                      e.target.value = e.target.value.replace(/[^0-9]/g, "");
-                      //  if length of input more than 12, display only 12 digits
-                      if (e.target.value.toString().length >= 12) {
-                        e.target.value = e.target.value.toString().slice(0, 12);
-                      }
-                    }}
-                    value={customTreeInputValue}
-                    type="text"
-                    inputMode="numeric"
-                    pattern="\d*"
-                    onChange={(e) => {
-                      setCustomTreeValue(e);
-                      setCustomTreeInputValue(e.target.value);
-                    }}
-                    ref={customInputRef}
-                  />
-                  <span>{t("trees")}</span>
-                </div>
-              </div>
-            </div>
+          <TaxDeductionOption />
 
-            {paymentSetup && paymentSetup.treeCost ? (
-              <p className="currency-selection mt-30">
-                <button
-                  onClick={() => {
-                    setopenCurrencyModal(true);
-                  }}
-                  className="text-bold text-primary"
-                  style={{ marginRight: "4px" }}
-                >
-                  {currency}{" "}
-                  <DownArrowIcon color={themeProperties.primaryColor} />
-                  {getFormatedCurrency(
-                    i18n.language,
-                    "",
-                    Number(paymentSetup.treeCost)
-                  )}{" "}
-                </button>
-                {t("perTree")}
-              </p>
-            ) : (
-              <div className={"mt-20"}>
-                <TreeCostLoader width={150} />
-              </div>
-            )}
+          <div className={"horizontal-line"} />
 
-            <TaxDeductionOption />
 
-            <div className={"horizontal-line"} />
-            {paymentSetup && paymentSetup.treeCost ? (
-              <div className={"w-100 text-center text-bold mt-20"}>
-                <span className={"text-primary"} style={{ marginRight: "4px" }}>
-                  {getFormatedCurrency(
-                    i18n.language,
-                    currency,
-                    paymentSetup.treeCost * treeCount
+          <DonationAmount/>
+
+          {paymentSetup && projectDetails ? (
+            minAmt && paymentSetup?.treeCost * treeCount >= minAmt ? (
+              !isPaymentOptionsLoading &&
+              paymentSetup?.gateways?.stripe?.account &&
+              currency ? (
+                <NativePay
+                  country={country}
+                  currency={currency}
+                  amount={formatAmountForStripe(
+                    paymentSetup?.treeCost * treeCount,
+                    currency.toLowerCase()
                   )}
-                </span>
-                {t("fortreeCountTrees", {
-                  count: Number(treeCount),
-                  treeCount: getFormattedNumber(
-                    i18n.language,
-                    Number(treeCount)
-                  ),
-                })}
-              </div>
-            ) : (
-              <div className={"text-center mt-20"}>
-                <TreeCostLoader width={150} />
-              </div>
-            )}
-
-            {paymentSetup && projectDetails ? (
-              minAmt && paymentSetup?.treeCost * treeCount >= minAmt ? (
-                !isPaymentOptionsLoading &&
-                paymentSetup?.gateways?.stripe?.account &&
-                currency ? (
-                  <NativePay
-                    country={country}
-                    currency={currency}
-                    amount={formatAmountForStripe(
-                      projectDetails.treeCost * treeCount,
-                      currency.toLowerCase()
-                    )}
-                    onPaymentFunction={onPaymentFunction}
-                    paymentSetup={paymentSetup}
-                    continueNext={() => setdonationStep(2)}
-                    isPaymentPage={false}
-                    paymentLabel={t("treesInCountry", {
-                      treeCount: treeCount,
-                      country: t(
-                        `country:${projectDetails.country.toLowerCase()}`
-                      ),
-                    })}
-                  />
-                ) : (
-                  <div className="mt-20 w-100">
-                    <ButtonLoader />
-                  </div>
-                )
+                  onPaymentFunction={onPaymentFunction}
+                  paymentSetup={paymentSetup}
+                  continueNext={() => setdonationStep(2)}
+                  isPaymentPage={false}
+                  paymentLabel={t("treesInCountry", {
+                    treeCount: treeCount,
+                    country: t(
+                      `country:${projectDetails.country.toLowerCase()}`
+                    ),
+                  })}
+                />
               ) : (
-                <p className={"text-danger mt-20 text-center"}>
-                  {t("minDonate")}{" "}
-                  <span>
-                    {getFormatedCurrency(i18n.language, currency, minAmt)}
-                  </span>
-                </p>
+                <div className="mt-20 w-100">
+                  <ButtonLoader />
+                </div>
               )
             ) : (
-              <div className="mt-20 w-100">
-                <ButtonLoader />
-              </div>
-            )}
-          </div>
+              <p className={"text-danger mt-20 text-center"}>
+                {t("minDonate")}{" "}
+                <span>
+                  {getFormatedCurrency(i18n.language, currency, minAmt)}
+                </span>
+              </p>
+            )
+          ) : (
+            <div className="mt-20 w-100">
+              <ButtonLoader />
+            </div>
+          )}
         </div>
       </div>
